ML OPS/dataops-final_wrk/5-ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
numpy==1.26.4
psycopg2-binary==2.9.9
prometheus-client==0.20.0
